refactor run to types
