+++ src/scranpy/dimensionality_reduction/types.py
