+++ src/scranpy/feature_selection/__init__.py
@@ -1,3 +1,3 @@
 from .choose_hvgs import ChooseHvgsOptions, choose_hvgs
 from .model_gene_variances import ModelGeneVariancesOptions, model_gene_variances
-from .run import FeatureSelectionStepOptions, FeatureSelectionStepResults
+from .types import FeatureSelectionStepOptions, FeatureSelectionStepResults
