+++ src/scranpy/marker_detection/__init__.py
@@ -1,2 +1,2 @@
-from .run import MarkerDetectionStepOptions, MarkerDetectionStepResults
 from .score_markers import ScoreMarkersOptions, score_markers
+from .types import MarkerDetectionStepOptions, MarkerDetectionStepResults
