+++ src/scranpy/nearest_neighbors/types.py
