+++ src/scranpy/normalization/__init__.py
@@ -1,2 +1,2 @@
 from .log_norm_counts import LogNormalizeCountsOptions, log_norm_counts
-from .run import NormalizationStepOptions, NormalizationStepResults
+from .types import NormalizationStepOptions, NormalizationStepResults
