+++ src/scranpy/quality_control/__init__.py
@@ -8,4 +8,4 @@ from .rna import (
     per_cell_rna_qc_metrics,
     suggest_rna_qc_filters,
 )
-from .run import RnaQualityControlOptions, RnaQualityControlResults
+from .types import RnaQualityControlOptions, RnaQualityControlResults
