+++ src/scranpy/quality_control/types.py
